Move fullscreen detection to WindowsGLRenderer
